feat_INT-34 add and сconfiguration logging -Выбрана и подключена logBack для логирования - Настроен шаблон для главного и других модулей
