+++ .gitignore
@@ -4,6 +4,7 @@ build/
 !**/src/main/**/build/
 !**/src/test/**/build/
 !**/src/
+logs/
 
 ### IntelliJ IDEA ###
 .idea/
