+++ .idea/gradle.xml
@@ -12,6 +12,7 @@
             <option value="$PROJECT_DIR$/client" />
             <option value="$PROJECT_DIR$/game" />
             <option value="$PROJECT_DIR$/server" />
+            <option value="$PROJECT_DIR$/untitled" />
           </set>
         </option>
       </GradleProjectSettings>
