+++ build.gradle.kts
@@ -31,6 +31,8 @@ allprojects {
         errorprone("com.google.errorprone:error_prone_core:2.28.0")
         // логирование
         implementation("ch.qos.logback:logback-core:1.5.6")
+        implementation("ch.qos.logback:logback-classic:1.5.6")
+        implementation("org.slf4j:slf4j-api:2.0.13")
         // кодогенерация
         compileOnly("org.projectlombok:lombok:1.18.32")
     }
