+++ client/src/main/java/io/deeplay/camp/ui/Cui.java
@@ -1,0 +1,3 @@
+package io.deeplay.camp.ui;
+
+public class Cui {}
