+++ client/src/main/java/io/deeplay/camp/ui/Gui.java
@@ -1,0 +1,3 @@
+package io.deeplay.camp.ui;
+
+public class Gui {}
