--- client/src/main/java/io/deeplay/ui/Cui.java
@@ -1,3 +1,0 @@
-package io.deeplay.ui;
-
-public class Cui {}
