--- game/src/main/java/io/deeplay/Main.java
@@ -1,11 +1,0 @@
-package io.deeplay;
-
-public class Main {
-  public static void main(String[] args) {
-    System.out.println("Hello, World!");
-    int x = 10;
-    for (int i = 0; i < x; i++) {
-      System.out.println("i = " + i);
-    }
-  }
-}
