+++ game/src/main/java/io/deeplay/camp/mechanics/GameEvent.java
@@ -1,0 +1,3 @@
+package io.deeplay.camp.mechanics;
+
+public interface GameEvent {}
