--- game/src/main/java/io/deeplay/mechanics/GameEvent.java
@@ -1,3 +1,0 @@
-package io.deeplay.mechanics;
-
-public interface GameEvent {}
