--- game/src/main/java/io/deeplay/mechanics/Move.java
@@ -1,3 +1,0 @@
-package io.deeplay.mechanics;
-
-public class Move {}
