--- game/src/main/java/io/deeplay/mechanics/Placement.java
@@ -1,3 +1,0 @@
-package io.deeplay.mechanics;
-
-public class Placement {}
