+++ server/src/main/java/io/deeplay/camp/controller/UserRequest.java
@@ -1,0 +1,3 @@
+package io.deeplay.camp.controller;
+
+public class UserRequest {}
