+++ server/src/main/java/io/deeplay/camp/dto/BoardResponseDto.java
@@ -1,3 +1,3 @@
-package io.deeplay.dto;
+package io.deeplay.camp.dto;
 
 public class BoardResponseDto {}
