+++ server/src/main/java/io/deeplay/camp/dto/PlacmentRequestDto.java
@@ -1,3 +1,3 @@
-package io.deeplay.dto;
+package io.deeplay.camp.dto;
 
 public class PlacmentRequestDto {}
