--- server/src/main/java/io/deeplay/controller/UserRequest.java
@@ -1,3 +1,0 @@
-package io.deeplay.controller;
-
-public class UserRequest {}
