+++ settings.gradle.kts
@@ -4,3 +4,4 @@ rootProject.name = "deeplay-camp-2024"
 include("client")
 include("server")
 include("game")
+include("untitled")
