--- src/main/java/io/deeplay/camp/Main.java
@@ -1,5 +1,0 @@
-package io.deeplay.camp;
-
-public class Main {
-  public static void main(String[] args) {}
-}
